+++ java/component/keyple-core/src/main/java/org/eclipse/keyple/core/seproxy/event/ObservableReader.java
@@ -177,27 +177,4 @@ public interface ObservableReader extends SeReader {
      */
     void setDefaultSelectionRequest(AbstractDefaultSelectionsRequest defaultSelectionsRequest,
             NotificationMode notificationMode, PollingMode pollingMode);
-
-    /**
-     * Signal sent by the application to the reader to indicate the end of the application
-     * processing.
-     * <p>
-     * Depending on whether a request with the indication CLOSE_AFTER has been executed before or
-     * not, a closing message will be sent to the reader in order to proceed with the closing of the
-     * physical channel.
-     * <p>
-     * The action to be continued will be the one defined by the PollingMode used to start the SE
-     * detection.
-     * <p>
-     * The call of this method is mandatory only if the current transaction performed via an
-     * observed reader did not end with a call to an operation causing a physical channel closing
-     * (CLOSE_AFTER). The main objective here is to achieve the closing of the physical channel. A
-     * typical use of this method is in handling exceptions (catch) that may occur during a
-     * transaction that would prevent reaching the last operation closing the physical channel.
-     * <p>
-     * The channel closing here refers to the application's intention to no longer communicate with
-     * the SE. This may take the form of an immediate physical channel closing in the case of an
-     * unobserved reader or a delayed closing upon removal of the SE (removal sequence).
-     */
-    void notifySeProcessed();
 }